Revert "chore: use size_t"
This reverts commit 80d4c67ea7736a2ec5bf263d85590a1686206f61.

+++ include/calc.h
@@ -1,17 +1,15 @@
 #ifndef CALC_H
 #define CALC_H
 
-#include <sys/types.h>
-
-static ssize_t
-max (ssize_t a, ssize_t b) {
+static int
+max (int a, int b) {
   __typeof__(a) _a = (a);
   __typeof__(b) _b = (b);
   return _a > _b ? _a : _b;
 }
 
-static ssize_t
-min (ssize_t a, ssize_t b) {
+static int
+min (int a, int b) {
   __typeof__(a) _a = (a);
   __typeof__(b) _b = (b);
   return _a < _b ? _a : _b;
