+++ include/xmalloc.h
@@ -4,7 +4,7 @@
 #include "exception.h"
 
 static inline void*
-xmalloc (size_t sz) {
+xmalloc (unsigned int sz) {
   void* ptr;
   if ((ptr = malloc(sz)) == NULL) {
     panic("[xmalloc::%s] failed to allocate memory\n", __func__);
